requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.4
Pillow==6.2.1
six==1.13.0
torch==1.3.1
torchvision==0.4.2
Werkzeug==0.16.0
